Move pingpong struct into its own package below testutil
This makes it clear that it is used only for tests, isolates it from
the rest of the Supervisor package, and allows it to be reused in other
tests as well.
